fix code warning

+++ window_scene/session/host/src/move_drag_controller.cpp
@@ -293,15 +293,15 @@ WSRect MoveDragController::CalcFreeformTargetRect(AreaType type, int32_t tranX, 
     FixTranslateByLimits(tranX, tranY);
     if (static_cast<uint32_t>(type) & static_cast<uint32_t>(AreaType::LEFT)) {
         targetRect.posX_ += tranX;
-        targetRect.width_ -= tranX;
+        targetRect.width_ -= static_cast<uint32_t>(tranX);
     } else if (static_cast<uint32_t>(type) & static_cast<uint32_t>(AreaType::RIGHT)) {
-        targetRect.width_ += tranX;
+        targetRect.width_ += static_cast<uint32_t>(tranX);
     }
     if (static_cast<uint32_t>(type) & static_cast<uint32_t>(AreaType::TOP)) {
         targetRect.posY_ += tranY;
-        targetRect.height_ -= tranY;
+        targetRect.height_ -= static_cast<uint32_t>(tranY);
     } else if (static_cast<uint32_t>(type) & static_cast<uint32_t>(AreaType::BOTTOM)) {
-        targetRect.height_ += tranY;
+        targetRect.height_ += static_cast<uint32_t>(tranY);
     }
     // check current ratio limits
     if (targetRect.height_ == 0) {
@@ -330,8 +330,8 @@ WSRect MoveDragController::CalcFixedAspectRatioTargetRect(AreaType type, int32_t
 {
     int32_t posX = originalRect.posX_;
     int32_t posY = originalRect.posY_;
-    int32_t width = originalRect.width_;
-    int32_t height = originalRect.height_;
+    int32_t width = static_cast<int32_t>(originalRect.width_);
+    int32_t height = static_cast<int32_t>(originalRect.height_);
     FixTranslateByLimits(tranX, tranY);
     if (mainMoveAxis_ == AxisType::UNDEFINED) {
         if (!InitMainAxis(type, tranX, tranY)) {
@@ -378,27 +378,27 @@ WSRect MoveDragController::CalcFixedAspectRatioTargetRect(AreaType type, int32_t
 void MoveDragController::CalcFreeformTranslateLimits(AreaType type)
 {
     if (static_cast<uint32_t>(type) & static_cast<uint32_t>(AreaType::LEFT)) {
-        minTranX_ = moveDragProperty_.originalRect_.width_ - limits_.maxWidth_;
-        maxTranX_ = moveDragProperty_.originalRect_.width_ - limits_.minWidth_;
+        minTranX_ = static_cast<int32_t>(moveDragProperty_.originalRect_.width_ - limits_.maxWidth_);
+        maxTranX_ = static_cast<int32_t>(moveDragProperty_.originalRect_.width_ - limits_.minWidth_);
     } else if (static_cast<uint32_t>(type) & static_cast<uint32_t>(AreaType::RIGHT)) {
-        minTranX_ = limits_.minWidth_ - moveDragProperty_.originalRect_.width_;
-        maxTranX_ = limits_.maxWidth_ - moveDragProperty_.originalRect_.width_;
+        minTranX_ = static_cast<int32_t>(limits_.minWidth_ - moveDragProperty_.originalRect_.width_);
+        maxTranX_ = static_cast<int32_t>(limits_.maxWidth_ - moveDragProperty_.originalRect_.width_);
     }
     if (static_cast<uint32_t>(type) & static_cast<uint32_t>(AreaType::TOP)) {
-        minTranY_ = moveDragProperty_.originalRect_.height_ - limits_.maxHeight_;
-        maxTranY_ = moveDragProperty_.originalRect_.height_ - limits_.minHeight_;
+        minTranY_ = static_cast<int32_t>(moveDragProperty_.originalRect_.height_ - limits_.maxHeight_);
+        maxTranY_ = static_cast<int32_t>(moveDragProperty_.originalRect_.height_ - limits_.minHeight_);
     } else if (static_cast<uint32_t>(type) & static_cast<uint32_t>(AreaType::BOTTOM)) {
-        minTranY_ = limits_.minHeight_ - moveDragProperty_.originalRect_.height_;
-        maxTranY_ = limits_.maxHeight_ - moveDragProperty_.originalRect_.height_;
+        minTranY_ = static_cast<int32_t>(limits_.minHeight_ - moveDragProperty_.originalRect_.height_);
+        maxTranY_ = static_cast<int32_t>(limits_.maxHeight_ - moveDragProperty_.originalRect_.height_);
     }
 }
 
 void MoveDragController::CalcFixedAspectRatioTranslateLimits(AreaType type, AxisType axis)
 {
-    int32_t minW = limits_.minWidth_;
-    int32_t maxW = limits_.maxWidth_;
-    int32_t minH = limits_.minHeight_;
-    int32_t maxH = limits_.maxHeight_;
+    int32_t minW = static_cast<int32_t>(limits_.minWidth_);
+    int32_t maxW = static_cast<int32_t>(limits_.maxWidth_);
+    int32_t minH = static_cast<int32_t>(limits_.minHeight_);
+    int32_t maxH = static_cast<int32_t>(limits_.maxHeight_);
     if (isDecorEnable_) {
         if (SessionUtils::ToLayoutWidth(minW, vpr_) < SessionUtils::ToLayoutHeight(minH, vpr_) * aspectRatio_) {
             minW = SessionUtils::ToWinWidth(SessionUtils::ToLayoutHeight(minH, vpr_) * aspectRatio_, vpr_);
@@ -425,19 +425,19 @@ void MoveDragController::CalcFixedAspectRatioTranslateLimits(AreaType type, Axis
 
     if (axis == AxisType::X_AXIS) {
         if (static_cast<uint32_t>(type) & static_cast<uint32_t>(AreaType::LEFT)) {
-            minTranX_ = moveDragProperty_.originalRect_.width_ - maxW;
-            maxTranX_ = moveDragProperty_.originalRect_.width_ - minW;
+            minTranX_ = static_cast<int32_t>(moveDragProperty_.originalRect_.width_) - maxW;
+            maxTranX_ = static_cast<int32_t>(moveDragProperty_.originalRect_.width_) - minW;
         } else if (static_cast<uint32_t>(type) & static_cast<uint32_t>(AreaType::RIGHT)) {
-            minTranX_ = minW - moveDragProperty_.originalRect_.width_;
-            maxTranX_ = maxW - moveDragProperty_.originalRect_.width_;
+            minTranX_ = minW - static_cast<int32_t>(moveDragProperty_.originalRect_.width_);
+            maxTranX_ = maxW - static_cast<int32_t>(moveDragProperty_.originalRect_.width_);
         }
     } else if (axis == AxisType::Y_AXIS) {
         if (static_cast<uint32_t>(type) & static_cast<uint32_t>(AreaType::TOP)) {
-            minTranY_ = moveDragProperty_.originalRect_.height_ - maxH;
-            maxTranY_ = moveDragProperty_.originalRect_.height_ - minH;
+            minTranY_ = static_cast<int32_t>(moveDragProperty_.originalRect_.height_) - maxH;
+            maxTranY_ = static_cast<int32_t>(moveDragProperty_.originalRect_.height_) - minH;
         } else if (static_cast<uint32_t>(type) & static_cast<uint32_t>(AreaType::BOTTOM)) {
-            minTranY_ = minH - moveDragProperty_.originalRect_.height_;
-            maxTranY_ = maxH - moveDragProperty_.originalRect_.height_;
+            minTranY_ = minH - static_cast<int32_t>(moveDragProperty_.originalRect_.height_);
+            maxTranY_ = maxH - static_cast<int32_t>(moveDragProperty_.originalRect_.height_);
         }
     }
 }
@@ -544,8 +544,8 @@ bool MoveDragController::IsPointInDragHotZone(int32_t startPointPosX, int32_t st
     Rect rectWithHotzone;
     rectWithHotzone.posX_ = winRect.posX_ - static_cast<int32_t>(HOTZONE_POINTER);
     rectWithHotzone.posY_ = winRect.posY_ - static_cast<int32_t>(HOTZONE_POINTER);
-    rectWithHotzone.width_ = winRect.width_ + static_cast<int32_t>(HOTZONE_POINTER) * 2; // double hotZone
-    rectWithHotzone.height_ = winRect.height_ + static_cast<int32_t>(HOTZONE_POINTER) * 2; // double hotZone
+    rectWithHotzone.width_ = winRect.width_ + HOTZONE_POINTER * 2u; // double hotZone
+    rectWithHotzone.height_ = winRect.height_ + HOTZONE_POINTER * 2u; // double hotZone
 
     if (sourceType == MMI::PointerEvent::SOURCE_TYPE_MOUSE &&
         !WindowHelper::IsPointInTargetRectWithBound(startPointPosX, startPointPosY, rectWithHotzone)) {
@@ -624,7 +624,7 @@ void MoveDragController::HandleMouseStyle(const std::shared_ptr<MMI::PointerEven
         winRect.posY_, winRect.width_, winRect.height_, newStyleID, oriStyleID);
     if (oriStyleID != newStyleID) {
         MMI::PointerStyle pointerStyle;
-        pointerStyle.id = newStyleID;
+        pointerStyle.id = static_cast<int32_t>(newStyleID);
         int32_t res = MMI::InputManager::GetInstance()->SetPointerStyle(0, pointerStyle);
         if (res != 0) {
             WLOGFE("set pointer style failed, res is %{public}u", res);
